Rubyficate git-cof
I do not like having multiple languages in a single source file so I figured I would convert this to Ruby. The main thing I wanted to do with this that would have been a real pain in Bash or AWK was to preserve the ordering of the branches that Git returns while injecting local branch names for remote branches that are yet to be checked out locally. Success.