s/blame/praise/g
"The project touched a LOT of files in the repo which I now realise was a huge mistake as my name will come up in `git blame`s for years to come 😂." --@shevaun

That project was incredible and the work that Shevaun led deserves praise. Now time to relearn that muscle-memory.